Fix bug that occurs when base page is only page and rework proxy rotation
When the base page was the only page, the scraper would throw an IndexError,
thinking that it was an IP Ban when it acutally wasn't.

Proxy rotation was reworked so that proxies are selected on random now, instead
of using the iteration cycle. This seems to keep the proxy list lasting longer
before getting banned.